refactor: 유저 파일 분리
