+++ src/main/java/io/spring/application/user/UserService.java
@@ -2,11 +2,6 @@ package io.spring.application.user;
 
 import io.spring.core.user.User;
 import io.spring.core.user.UserRepository;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import javax.validation.Constraint;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -53,54 +48,5 @@ public class UserService {
         updateUserParam.bio(),
         updateUserParam.image());
     userRepository.save(user);
-  }
-}
-
-@Constraint(validatedBy = UpdateUserValidator.class)
-@Retention(RetentionPolicy.RUNTIME)
-@interface UpdateUserConstraint {
-
-  String message() default "invalid update param";
-
-  Class[] groups() default {};
-
-  Class[] payload() default {};
-}
-
-class UpdateUserValidator implements ConstraintValidator<UpdateUserConstraint, UpdateUserCommand> {
-
-  @Autowired private UserRepository userRepository;
-
-  @Override
-  public boolean isValid(UpdateUserCommand value, ConstraintValidatorContext context) {
-    String inputEmail = value.getParam().email();
-    String inputUsername = value.getParam().username();
-    final User targetUser = value.getTargetUser();
-
-    boolean isEmailValid =
-        userRepository.findByEmail(inputEmail).map(user -> user.equals(targetUser)).orElse(true);
-    boolean isUsernameValid =
-        userRepository
-            .findByUsername(inputUsername)
-            .map(user -> user.equals(targetUser))
-            .orElse(true);
-    if (isEmailValid && isUsernameValid) {
-      return true;
-    } else {
-      context.disableDefaultConstraintViolation();
-      if (!isEmailValid) {
-        context
-            .buildConstraintViolationWithTemplate("email already exist")
-            .addPropertyNode("email")
-            .addConstraintViolation();
-      }
-      if (!isUsernameValid) {
-        context
-            .buildConstraintViolationWithTemplate("username already exist")
-            .addPropertyNode("username")
-            .addConstraintViolation();
-      }
-      return false;
-    }
   }
 }
